nix: build example packages
This commit refactors the Nix flake to build the example packages,
further validating that the package works as intended. It adds these
builds to CI and vendors the dependencies so that the vendor hash
doesn't need to be manually bumped whenever dependencies are bumped by
dependabot.

A future PR should add tests that exercise the generated PAM modules
using `pamtester`.

Signed-off-by: squat <lserven@gmail.com>